+++ predictsignauxfaibles/__main__.py
@@ -11,8 +11,8 @@ import pandas as pd
 
 from predictsignauxfaibles.config import OUTPUT_FOLDER, IGNORE_NA
 from predictsignauxfaibles.data import SFDataset
-from predictsignauxfaibles.explainability import explain
 from predictsignauxfaibles.evaluate import evaluate
+from predictsignauxfaibles.explain import explain
 from predictsignauxfaibles.pipelines import run_pipeline
 from predictsignauxfaibles.utils import (
     set_if_not_none,
